Add unit tests for each pull request targeting the dot-rocMLIR branch
Enable TRITON_USE_ROCM by default

In the current logic, TRITON_USE_ROCM is on only when HIP is found on
the system. However, HIP is usually not found on runner images on
github.
We need this macro since the mma* behavior is disabled only when
USE_ROCM is defined. And rocMLIR path requires mma* path to be disabled.