Fix test_convert2d
In the recent IFU, the mask of store op was changed to something like
mask = tid * elemsPerThread < numElems
where elemsPerThread is the total number of elements each thread needs
to take care of, which is computed out of the tensor shape of the
layout encoding, and numElems is the total number of elements in the
tensor.
This is not technically correct since each thread is not storing
`elemsPerThread` elements contiguously. And this is caught when
the CTA coverage, obtained from `getShapePerCTA`, cannot be covered
by the tensor, which is the case with layout0 in test_convert2d.