Fixed the issue with insert_slice op
Previously, the insert_slice op comes from the tensor dialect. Its
offset must be an index. However, when insert_slice_async is
decomposed into load and insert_slice, an I32 value is passed to the
insert_slice's builder as its offset.

> Why isn't this an issue for upstream?
- Because the insert_slice op is lowered to LLVMIR right after its
creation in the same pass. Therefore, no checker/verifier/sanitizer
are aware of this mismatch issue.
- the decompose of insert_slice_async into load + insert_slice happens
in a walker rather than a pattern rewriter. The generated insert_slice
is also not checked.

> The solution
We borrowed tensor.insert_slice into triton_gpu.insert_slice as triton
does for extract_slice op. The offsets (and also sizes and strides)
are set to I32.

> TODO
Since insert_slice op is copied from the tensor dialect, we may not
depend on the TensorDialet anymore.