Support chained dot in test_dot (without k-loop)
The missing piece is the emitOffsetForMfmaLayout() function used by
emitIndices() to compute the offset of each element for the current
thread. This is required by convert_layout #mfma -> #lds, which is
present in chained-dot case but not in single-dot case.
In chain-dot case, the result of the first dot (#mfma) is used as one
of the operands (#dot_op) of the second dot. Therefore, a
convert_layout is inserted to convert #mfma to #dot_op, which is
decomposed to
convert_layout #mfma -> #lds and
convert_layout #lds -> #dot_op