Ignore
This is a combination of 3 commits.

Minimize diff

This is a combination of 3 commits.

This is a combination of 2 commits.

Compress

This is a combination of 4 commits.

Compress work

This is a combination of 2 commits.

Compress work

This is a combination of 3 commits.

Compress work

This is a combination of 12 commits.

remove old scripts

add scripts back

ignore stuff

ignore csv and png

disable the other benches

just run one step

report ms

clean

save

just run empty kernel

ignore

save

min diff

must comment out everythings else when testing empty kernel

min diff

use 128

save the different config

save

use the autotune

save

save

save

save